CVXR namespace issues
The package relies on functionality exported from CVXR, but its namespace is not
attached. Even if the user runs `library(CVXR)` before using the code, they will
encounter an error since the S4 generic of `solve` will not be available inside
of the package's function closures.

We can fix this without altering the namespace by prefixing CVXR-related
functionality with `CVXR::`